+++ NEWS.md
@@ -1,3 +1,7 @@
+# spectralAnomaly (development version)
+
+* Add new function to create artificial anomalies to a vector.
+
 # spectralAnomaly 0.1.1
 
 * DESCRIPTION file corrections